Add a bound in the workflow that generalized-bulletproofs compiles with Rust 1.69
This effectively changes our MSRV policy by somewhat committing to a Rust 1.69
target for this library. If in the future, we have a PR which causes this
action to fail, we'd have to decide whether we want to enforce it as a
requirement or remove the bound.

Simultaneously performs the `cargo update`s required to achieve this MSRV, and
updates the `generalized-bulletproofs` Cargo.toml to require the necessary
dependency versions. It already specified the same minor versions, but these
patch versions are necessary for the MSRV to be satisfied.