Extensively introduce internal bounds
Resolves https://github.com/monero-oxide/monero-oxide/issues/72.

This commit's most notable changes are to `monero-io`, which has been widened
to accept `FnMut` over `Fn` (as all `Fn` are `FnMut`), and to the handling of
arbitrary data as defined in `monero-wallet`.

Prior, `monero-wallet` would only create transactions whose extra was less than
the limit imposed by a relay limit, yet would accept arbitrary data from an
extra of any size. Now, when parsing arbitrary data out of a transaction, only
the arbitrary data fields within the first 1060 bytes are considered. This lets
us establish a bound on the amount of arbitrary data despite extra itself being
unbounded. The serialization of `Metadata` has also been adjusted to use a
`VarInt` over a `u64`.

The bump to the version of `std-shims` is required for
`impl<R: Read> Read for &mut R`, which is used inside the new
`Transaction::read` function (which now enforces a maximum byte length on
non-miner transactions).

Split out of https://github.com/monero-oxide/monero-oxide/pull/66.
Specifically, this represents the following commits:

commit e041ad78e1c209ba5a66e7f62f3bd87ae650093e
Fix `no-std` builds

---

commit 2ab2db1bce641f1610dd7729fdcc4542d8b7f55d
Update citation for the TX extra relay rule now that it's merged

Also adds a test for the new limit on arbitrary data, and prefers `_SIZE` to
`_LEN` in constants I've declared.

---

commit 57d5bca53f2ec710539ee790cd8a7798669a07df
Bound the size of addresses when deserializing a `SignableTransaction`

---

commit 63cc9361403bc9b6d7837eea531fe208894154d4
Introduce length bound on the size of an address string with `SignableTransaction` deserialization

---

commit 1eebe1949d9af58971ac0a92d106ceb747cb61a3
Rename `MAX_INPUTS` to `INPUTS_UPPER_BOUND`, as it isn't accurate

Also uses it within `SignableTransaction` deserialization.

---

commit 3f03b0f9a8042ac2f58cfcd52166a8768c7b852e
Introduce bound to `Decoys` during deserialization

---

commit 123288e509d41a36940f2249b2b1d97b09a94285
Bound the total amount of arbitrary data to the relay rule for the maximum size for a transaction's extra

Changes the encoding of `arbitrary_data` within `Metadata` to use a varint for
a prefix, not a `u64`.

---

commit f364cfac53e95264d3d111ea76875dcdca1b3d30
Add `MAX_NON_MINER_TRANSACTION_SIZE` and `MAX_INPUTS`

Uses these bounds within `read_vec` calls in the transaction module where
possible.

---

commit b1e6dc030662ba6e9418a5216332e95c7701c852
Widen `Fn` bounds to `FnMut` in `monero-io`

This allows more flexibility to callers and works as all `Fn` are also `FnMut`.