Incorporate the inputs' commitments into the seeded RNGs for proving Bulletproofs, deriving pseudo-out masks
The seeded RNGs are bound to the outgoing view key and the inputs' keys. As the
input's keys are derived from a shared key, so should be the commitments,
meaning the RNG seed was already binding to the commitments under the Monero
wallet protocol, for a given address. In order to make this clearer, the
commitments are now bound in the seeded RNG transcript, ensuring the lack of
reuse of any ephemeral secrets within the Bulletproof prover.

The same idea is applied to the derivation of pseudo-out masks when producing
multisig transactions, as two outputs with the same keys yet different amounts
would have pseudo-outs with the same masks yet different amounts, allowing
recovering the difference of their value.

All of this _should_ be unnecessary as these commitments were _already_ part
of the transcript for outputs derived via the Monero protocol. This just
straightens things out.