Move `monero-oxide/verify-chain` to `tests/verify-chain` for clarity
We have bash scripts to iterate over all members within the workspace. They
attempt to remove the trailing comma from a toml array to make it a valid JSON
array. The existing bash script failed to consider the line we remove the
trailing comma from may itself be removed, shifting which line is the trailing
comma. Here, we don't run the unit tests for `verify-chain` (by removing the
line for it), yet it's now the last line in the array. Accordingly, the bash
scripts to iterate over all members of the workspace have been updated to
handle this edge case properly (as required). I just synchronized them with
updates I made within the Serai repository, which also added support for
removing comments.