Better document arbitrary data
`monero-wallet` already documented a byte _custom to `monero-wallet`_ was used.
Accordingly, while arbitrary data was a fingerprint (as it fundamentally is),
it was a _documented_ fingerprint.

However, despite already saying it was a marker custom to `monero-wallet`, we
could have better documented the method it was embedded with and the
considerations relevant.

This clearly details the protocol and puts forth a RECOMMENDED format for any
arbitrary data (padded, indistinguishable from random).