Remove methods to fetch a batch of blocks by their hashes
This method requests the interface perform an unordered read, which is likely
to be significantly slower than an ordered read. The fact this method existed
next to the methods which provided ordered reads, and would be signficantly
faster, suggested it as an alternative (which was a performance footgun).

The only benefit lost is prior, we would only submit one literal request (one
HTTP packet) and now we'd send `n`. This better models the linear cost of
performing these requests however.