Implement `ProvidesUnvalidatedBlockchain::contiguous_blocks` without relying on `get_blocks.bin`
This prevents unnecessarily downloading the transactions when we want the
blocks alone.

It should be implemented with a batch request, and the current code attempts to
use one, yet Monero's RPC server is incomplete and does not implement the
JSON-RPC 2.0 specification properly. If the server fails to decode the batch
request, as Monero's RPC server does, we assume the server doesn't support
batch requests and will linearly execute each of the requests.

https://github.com/monero-project/monero/issues/10118 to track support for
batch RPC requests within Monero.