Allow ':' within passwords, '@' within URLs
These are both allowed per https://www.rfc-editor.org/rfc/rfc3986.

This raises an example of the issue with having our own URL handling here. On
the one hand, we have only about 50 lines of code and this works for most
cases, but obviously this doesn't attempt to match any specific specification
and has inconsistencies with presumably all modern specifications. It just
happens to work in most cases for just 50 lines of code.

We could adopt `url`, yet I wasn't happy with how many dependencies it had in
relation to `simple-request`. We could not accept a URL, yet rather an explicit
`scheme, username, password, domain, port, path_query_and_fragment`? That would
defer URL parsing to the user (or the lack-of if they can propagate taking
these arguments individually).