Updates to the Merkle tree (#104)
* Update `merkle_root` to not perform any allocations

Prior, the function would allocate for the new layer of the tree, causing
memory consumption to increase by up to 50% of the size of the input. This is
unecessary, as demonstrated by this commit.

It'd also perform a bunch of small allocations (64 bytes) for every single
pairing, which is now replaced with 64-bytes on the stack.

This isn't performance critical code, so this should be minor, but it's good to
have done. I also increased the documentation. Hopefully, with FCMP++, we can
enforce no transactions are less than 64 bytes so long as every miner
transaction had an extra of >20 bytes or so? (as CARROT will make each output
larger than 64 bytes)

This also updates the calculation of `low_pow_2` in a way which hanndles if
`leaves.len()`'s MSB is set, superseding my commits in
https://github.com/monero-oxide/monero-oxide/pull/94.

* Add test for and document the properties of the preimage for the hash defined for block 202,612

* Add disclaimer to `serialize_pow_hash` regarding block #202,612