Add `monero-ed25519` (partial)
This was discussed in https://github.com/monero-oxide/monero-oxide/issues/64.
The most recent Cuprate meeting also mentioned this change publicly, with
boog900 publicly acknowledging their support. The motivation is to allow us to
upgrade to `curve25519-dalek 5.0` without it being a breaking change, or even
other implementations entirely.

This updates `io`, `generators`, `primitives`, `borromean`, `bulletproofs`,
`mlsag`, and `address`. `ed25519`, `io`, `primitives`, `borromean`,
`bulletproofs`, `mlsag` no longer have `curve25519-dalek` in their public API.

`clsag`, `oxide`, `rpc`, and `wallet` have yet to be updated. I believe we'll
refrain from removing `curve25519-dalek` from the `rpc`, `wallet` APIs at this
time as they aren't being targeted for stabilization and users do require misc.
key generation functionality.

`generators` was updated yet will likely be removed by the end of this PR, or
at least moved to `bulletproofs-generators`. `primitives` now only contains
`Decoys`, `UpperBound`, where we should be able to move `Decoys` to `clsag`?
This is as we don't support proving non-CLSAG proofs and the next upgrade will
be FCMPs, so we don't need to maintain this as a generic primitive. This raises
the question of if we want to keep `monero-primitives` with `UpperBound` or
remove it entirely.

This commit also resolves
https://github.com/monero-oxide/monero-oxide/issues/124.