Remove allocations from `monero_base58::{decode, decode_check}`
This replaces the `monero-primitives` dependency with `sha3` directly, which I
don't love, but it's necessary and not committed to within our public API, so
it should be fine.

This also ensures, when decoding an address, the input isn't longer than the
bound, ensuring we won't allocate for the decoded data if the input is of
absurd length. While this wouldn't be an issue, as the allocation would be
smaller than the existing memory used for the input, it is a bit safer.