Use `MaybeUninit` for the large arrays written into
This seems to save ~10% which is very worthwhile. Unfortunately, the current
version technically is immediate UB. Each element in the array must be
`MaybeUninit`, written to, then the entire array must be transmuted (not as
seen here).

Alternatively, we could look at a manual memory allocation to avoid overhead of
Rust's initialization, range checks.