Represent `WCG` not as a non-sparse `Vec` of sparse vectors, but a sparse `BTreeMap` of sparse vectors
Within the Monero protocol, the non-sparse `Vec` was spending hundreds of MB on
empty vectors (just 16 bytes each) for commitments which had no terms
constrained.

An earlier attempt tried a truly sparse vector, yet that invoked a O(n^2)
algorithm to compute as the vector had to be iterated to find all entries
within a specific vector, causing an O(n) invocation per commitment. This
tripled the verification time.

This algorithm represents the memory in O(n log c) where n is the
constrained terms and c is the amount of commitments which have constrained
terms. While the amount of commitments is large, the amount constrained within
a single `LinComb` should be small. Processing is an `O(log c)` lookup before
the traditional `O(n)` iteration. Accordingly, this should have almost all of
the memory savings with almost none of the performance cost.

cc @j-berman to provide exact performance numbers for the 1-in, 2-in, 4-in,
8-in, and 128-in cases (verification and time to verify) on
https://github.com/monero-oxide/monero-oxide/pull/39. cc @boog900 to review
this change to ensure it doesn't change the wires of the encoded program.