Introduce a `const fn` hash-to-point
Removes the `monero-bulletproofs-generators` crate and `monero-bulletproofs`
`build.rs`. This is solely intended as an experiment and unfortunately cannot
be merged without significant algorithmic improvements to it itself or Rust's
`const fn` evaluation.

Removal of `batch_sqrt` for the addition chain
(https://github.com/monero-oxide/monero-oxide/issues/131) would have similar
computational requirements with a fraction of the memory, and may be
worthwhile.