Store up to four values on the stack before opening a new allocation within a `LinComb`
This increases the memory usage (as empty/small vectors now have a minimum size
to them), but reduces the amount of allocations. To offset the performance
impact, the specialized `extend_from_slice` has been used where applicable.
This is without consequence EXCEPT to the member of list gadget which would
have had to perform a clone. It's been optimized to negate once at the start of
the gadget so we may use traditional borrows in every case after.